requirements.txt
Django
djangorestframework
psycopg2
pillow
requests
python-decouple==3.8
base64